fix: move language and theme providers to its own components
