+++ src/shared/lang-toggle/lang-toggle.stories.tsx
@@ -1,7 +1,8 @@
-import '../../i18n/config';
-import { Meta, StoryObj } from '@storybook/react';
-import LanguageToggle, { LanguageProvider } from './lang-toggle';
 import React from 'react';
+import { Meta, StoryObj } from '@storybook/react';
+import '../../i18n/config';
+import LanguageProvider from '../lang-provider/lang-provider';
+import LanguageToggle  from './lang-toggle';
 import { useTranslation } from 'react-i18next';
 
 const meta: Meta<typeof LanguageToggle> = {
