+++ src/shared/theme-provider/theme-provider.tsx
@@ -1,0 +1,34 @@
+import React, { FC, ReactNode, createContext, useContext, useState } from 'react';
+
+type Theme = "light" | "dark";
+
+interface ThemeContextType {
+    theme: Theme;
+    toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType>({
+    theme: "light",
+    toggleTheme: () => {},
+});
+
+type ThemeProviderProps = {
+    children: ReactNode;
+};
+
+/**
+ * Провайдер переключения тем
+ */
+export const ThemeProvider = ({ children }: ThemeProviderProps) => {
+    const [theme, setTheme] = useState<Theme>("light");
+
+    const toggleTheme = () => {
+        setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+    };
+
+    return (
+        <ThemeContext.Provider value={{ theme, toggleTheme }}>
+            {children}
+        </ThemeContext.Provider>
+    );
+};
