+++ src/shared/theme-toggle/theme-toggle.tsx
@@ -1,36 +1,8 @@
-import React, { FC, ReactNode, createContext, useContext, useState } from 'react';
+import React, { FC, useContext } from 'react';
+import { ThemeContext } from '../theme-provider/theme-provider';
 import cn from 'clsx';
 import './theme-toggle.css';
 
-type Theme = "light" | "dark";
-
-interface ThemeContextType {
-    theme: Theme;
-    toggleTheme: () => void;
-}
-
-export const ThemeContext = createContext<ThemeContextType>({
-    theme: "light",
-    toggleTheme: () => {},
-});
-
-type ThemeProviderProps = {
-    children: ReactNode;
-};
-
-export const ThemeProvider = ({ children }: ThemeProviderProps) => {
-    const [theme, setTheme] = useState<Theme>("light");
-
-    const toggleTheme = () => {
-        setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-    };
-
-    return (
-        <ThemeContext.Provider value={{ theme, toggleTheme }}>
-            {children}
-        </ThemeContext.Provider>
-    );
-};
 
 /**
  * Компонент переключения тем
